refactor: 프로필 변경 로직 service로 분리
