+++ src/domains/mypage/service/index.ts
@@ -1,0 +1,1 @@
+export { default as useEditProfileService } from "./useEditProfileService";
